Added the ability to reduce the will of prisoners by beating. This works in the same way as reducing resistance - a prisoner with the interaction set to 'reduce will' will be occasionally beaten by breakers. This will result in a proportional decrease in their will using the same calculations as for reducing resistance.
Changed some labels and descriptions to reflect the new functionality. Since enslaving prisoners requires the Ideology DLC, the relevant texts are only visible when it's active.

Defined a new set of words used to generate events in the 'combat log' during the beating, which replaced 'ConceptBadass' as it contained words that did not fit the situation. It's basically the same as 'ConceptBadass', only without those words and with some minor additions.

Updated translations using DeepL.